+++ lizard_measure/templates/portals/doelen-beheer.js
@@ -1,7 +1,6 @@
 {% load get_grid %}
 {% load get_portal_template %}
 
-{% if perms.auth.is_analyst %}
 
 {
     itemId: 'doelen-beheer',
@@ -57,6 +56,4 @@
         }]
 	}]
 }
-{% else %}
-    {% get_portal_template geen_toegang %}
-{% endif %}
+