+++ lizard_measure/templates/portals/maatregelen-beheer.js
@@ -1,7 +1,6 @@
 {% load get_grid %}
 {% load get_portal_template %}
 
-{% if perms.auth.is_analyst %}
 
 {
     itemId: 'maatregelen-beheer',
@@ -95,12 +94,15 @@
                 {name: 'id', title: 'id', editable: false, visible: false, width: 30, type: 'number'},
                 {name: 'ident', title: 'ident', editable: false, visible: false, width: 100, type: 'text'},
                 {name: 'title', title: 'titel', editable: true, visible: true, width: 200, type: 'text'},
+                {name: 'import_source', title: 'bron', editable: false, sortable: false, visible: true, width: 100, type: 'text'},
+                {name: 'status_planned', title: 'planning', editable: false, visible: true, sortable: false, width: 100, type: 'text'},
+                {name: 'status_realisation', title: 'realisatie', editable: false, sortable: false, visible: true, width: 100, type: 'text'},
                 {name: 'is_KRW_measure', title: 'KRW maatregel', editable: true, visible: true, width: 100, type: 'boolean'},//automatisch aanmaken
                 {name: 'is_indicator', title: 'focus maatregel', editable: true, visible: true, width: 100, type: 'boolean'},
                 {name: 'parent', title: 'onderdeel van', editable: false, visible: true, width: 75, type: 'combo'},
                 {name: 'measure_type', title: 'maatregeltype', editable: true, visible: true, width: 150, type: 'combo', choices: Ext.JSON.decode({% autoescape off %}'{{ measure_types }}'{% endautoescape %})},//todo: voeg choices to vanuit model
                 {name: 'period', title: 'periode', editable: true, visible: true, width: 100, type: 'combo', choices: Ext.JSON.decode({% autoescape off %}'{{ periods }}'{% endautoescape %})},//todo: voeg choices to vanuit model
-                {name: 'categories', title: 'beleidsdoel', editable: false, visible: true, width: 150, type: 'combo', multiSelect: true, choices: Ext.JSON.decode({% autoescape off %}'{{ categories }}'{% endautoescape %})},
+                {name: 'categories', title: 'beleidsdoel', editable: false, visible: true, sortable: false, width: 150, type: 'combo', multiSelect: true, choices: Ext.JSON.decode({% autoescape off %}'{{ categories }}'{% endautoescape %})},
                 {name: 'value', title: 'waarde', editable: true, visible: true, width: 75, type: 'number'},
                 {name: 'unit', title: 'eenheid', editable: true, visible: true, width: 75, type: 'combo', choices: Ext.JSON.decode({% autoescape off %}'{{ units }}'{% endautoescape %})},
                 {name: 'initiator', title: 'initiatiefnemer', editable: true, visible: true, width: 100,
@@ -131,16 +133,17 @@
                 },
                 {name: 'responsible_department', title: 'afdeling', editable: true, visible: true, width: 75, type: 'text'},
 
-                {name: 'total_costs', title: 'totale kosten', editable: false, visible: true, width: 75, type: 'number'},
+                {name: 'total_costs', title: 'totale kosten', editable: true, sortable: false, visible: true, width: 75, type: 'number'},
                 {name: 'investment_costs', title: 'investeringskosten', editable: true, visible: true, width: 75, type: 'number'},
                 {name: 'exploitation_costs', title: 'exploitatiekosten', editable: true, visible: true, width: 75, type: 'number'},
+                {name: 'land_costs', title: 'grondkosten', editable: true, visible: true, width: 75, type: 'number'},
+                {name: 'target_esf', title: 'Doel ESF', editable: false, sortable: false, visible: true, width: 100, type: 'text'},
+                {name: 'effect_esf', title: 'Effect op ESF', editable: false, sortable: false, visible: true, width: 100, type: 'text'},
 
-                {name: 'waterbodies', title: 'KRW waterlichamen', editable: false, visible: true, width: 100, type: 'gridcombobox'},
-                {name: 'areas', title: 'Aan/ afvoergebieden', editable: false, visible: true, width: 100, type: 'gridcombobox'}
+                {name: 'waterbodies', title: 'KRW waterlichamen', editable: false, sortable: false, visible: true, width: 100, type: 'gridcombobox'},
+                {name: 'areas', title: 'Aan/ afvoergebieden', editable: false, sortable: false, visible: true, width: 100, type: 'gridcombobox'}
            ]
         }]
 	}]
 }
-{% else %}
-    {% get_portal_template geen_toegang %}
-{% endif %}
+